Replaced array of threads with a Stack and LinkedList
threads of type Stack is the empty thread pool. When server is connected
it first tests to see if the stack is empty. If it is it gives server
full message. If it is not empty it pops a thread out, initializes it,
adds it to is running_threads LinkedList, and runs it. When thread is
done running it removes thread from running_threads LinkedList and
pushes on a new null pointing thread.